xf86-video-dummy: changeable physical output size
Add WIDTH_MM and HEIGHT_MM output properties to allow setting "physical"
size of the output. Normally this data is retrieved from EDID, but for
obvious reasons it can't be done in the VM.
This patch allows to set this properties from user space. This is later
reported as output size (the header line describing the output in
`xrandr` output).

Signed-off-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>